+++ src/components/index.js
@@ -4,3 +4,4 @@ export { SearchBar } from "./SearchBar/SearchBar";
 export { SideBar } from "./SideBar/SideBar";
 export { FilterModal } from "../components/Modals/FilterModal/Filtermodal.jsx";
 export { ColorPallete } from "../components/Modals/ColorPalleteModal/ColorPallete.jsx";
+export { LabelModal } from "../components/Modals/LabelModal/LabelModal.jsx";
