+++ src/context/index.js
@@ -1,0 +1,5 @@
+export { useAuth } from "../context/authContext/authenticationContext.jsx";
+export { useLabel } from "../context/labelContext/LabelContext.jsx";
+export { useTheme } from "../context/noteThemeContext/noteThemeContext";
+export { usePriority } from "../context/priorityContext/priorityContext.jsx";
+export { useServices } from "../context/serviceContext/serviceContext.jsx";
